# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Companies ────────────────────────────────────────────────────
...
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List deals
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: ownerId
          in: query
          schema:
            type: integer
        - name: contactId
          in: query
          schema:
            type: integer
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create a deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    get:
      operationId: getDeal
      tags: [deals]
      summary: Get a deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    patch:
      operationId: updateDeal
      tags: [deals]
      summary: Update a deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    delete:
      operationId: deleteDeal
      tags: [deals]
      summary: Delete a deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Tasks ────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ── Auth ───────────────────────────────────────────────────
...
    Contact:
      type: object
      required: [id, firstName, lastName, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        status:
          type: string
          enum: [lead, prospect, customer, churned, inactive]
        source:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ContactInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        title:
          type: string
        companyId:
          type: integer
        status:
          type: string
          enum: [lead, prospect, customer, churned, inactive]
        source:
          type: string
        tags:
          type: string
        notes:
          type: string
        ownerId:
          type: integer
    ContactUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        title:
          type: string
        companyId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [lead, prospect, customer, churned, inactive]
        source:
          type: string
        tags:
          type: string
        notes:
          type: string
        ownerId:
          type: ["integer", "null"]

    # ── Deals ──────────────────────────────────────────────────
    Deal:
      type: object
      required: [id, title, stage, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        stage:
          type: string
          enum: [prospecting, qualification, proposal, negotiation, closed_won, closed_lost]
        value:
          type: ["number", "null"]
        currency:
          type: string
        probability:
          type: ["integer", "null"]
        expectedCloseDate:
          type: ["string", "null"]
          format: date
        contactId:
          type: ["integer", "null"]
        contactName:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DealInput:
      type: object
      required: [title, stage]
      properties:
        title:
          type: string
        stage:
          type: string
          enum: [prospecting, qualification, proposal, negotiation, closed_won, closed_lost]
        value:
          type: number
        currency:
          type: string
        probability:
          type: integer
        expectedCloseDate:
          type: string
          format: date
        contactId:
          type: integer
        companyId:
          type: integer
        ownerId:
          type: integer
        notes:
          type: string
    DealUpdate:
      type: object
      properties:
        title:
          type: string
        stage:
          type: string
          enum: [prospecting, qualification, proposal, negotiation, closed_won, closed_lost]
        value:
          type: ["number", "null"]
        currency:
          type: string
        probability:
          type: ["integer", "null"]
        expectedCloseDate:
          type: ["string", "null"]
          format: date
        contactId:
          type: ["integer", "null"]
        companyId:
          type: ["integer", "null"]
        ownerId:
          type: ["integer", "null"]
        notes:
          type: string

    # ── Tasks ──────────────────────────────────────────────────
...